requirements.txt
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
certifi==2026.2.25
-e git+https://github.com/eseoh69/Datasheet-ai.git@ebf5f39a1a4cf0a888598bc7d951fe55f576d3bf#egg=datasheet_ai
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
